server/workers/scraper_worker/requirements.txt
playwright==1.47.0
google-cloud-firestore>=2.16.0
httpx>=0.27.0
python-dotenv>=1.0.1
